refactor(tenant/wizard): split and simplify property creation wizard
- Extract wizard configuration into wizard-config.ts
- Extract navigation logic into use-wizard-navigation.ts
- Consolidate layout exports into wizard-layout.tsx
- Add wizard-step-renderer.tsx and update imports
- Remove duplicate/unused layout files
- No functional changes; ran TS checks
